* don't start transaction on all writes (but do set sticky on) * automatically switch database of backends when applied to new clients * remove some dead code
We work again, should be safe to use.  Would be interesting to see if there
any issues.  (Not that I really think anybody but me is using this atm,
but still...)

I don't think writes are all properly tagged yet... and I'm certain there
are other issues undiscovered.  But queries are going through.