* add proxy->server authentication
proxy_user and proxy_pass are used to set the auth credentials sent to the server.
They will do the right thing and authenticate against the server.

client->proxy authentication still does not work, anybody who can connect to the
ip:port of the proxy can use the configured backends.